Refactor vatsim routes to separate files
